+++ src/tripCard/TripCard.style.tsx
@@ -53,10 +53,8 @@ export const StyledTripCard = styled(Card)`
 
   & .kirk-tripCard-itinerary {
     min-width: 0; /* reset flex default to allow shrinking below content width */
-  }
-
-  &.kirk-tripCard--with-price .kirk-tripCard-itinerary {
-    padding-right: ${space.m};
+    position: relative;
+    left: -${space.xl}; // Counters normalization of Itinerary component
   }
 
   & .kirk-tripCard-price {